Git projects are loaded into fzf "asynchronously"
`fzf` is now loaded with local projects, and when the github projects
are loaded they are added to fzf. The downside with this approach is
that the github projects must be loaded before the program can launch a
project, however the program feels way more responsive because the
window is up faster. The time for the github projects to get fetched
also is slightly masked by the user typing and selecting a project.

- Created fzf_intergration module
- Removed `fzf_rust`, however I ended up modifying the code
- Documented functions