feat!: Add configuration option for multiple windows in tmux
Users can now define multiple windows to start in their tmux sessions
for their projects using their configuration file. Written unit tests to
verify that the produced config is as expected. Next step is to handle
what to do if the number of windows doesn't match the number of start
commands

BREAKING CHANGE: The configuration option for `window_name` has been
renamed to `window_names` and is now an array of strings, and the
configuration option for `on_open` has been renamed to `start_commands`
and is also an array of strings