feat!: Add config option for choosing any fzf layout
Users can now configure fzf to use any of the three fzf layouts using
the `layout` option under the `fzf` table. The configuration as always
is best effort, so if the option is invalid the default is chosen

BREAKING CHANGE: replaced `reverse_layout` fzf configuration option for
the `layout` configuration option.