fix: Fix cloning remote repos
Fixed borked cloning of repos from the open menu. I broke it when I
added the git status prepending, however now it works again. Essentially
it looks through the vec of repos first with the raw output of fzf, and
if it can't find it, it'll take a slice minus the length of the project
indicator prepend, and search again.