feat: add `enabled` configuration option for tmuxinator integration (#20)
If `true`, then tmuxinator integration is used as currently implemented. If `false`, then a named tmux session in the project root is created instead of running tmuxinator. Useful for speed improvements, as well as one less program required.

Defaults to `true`

Planning to expand on this later to hopefully remove the need for tmuxinator, as it's function is something that should be handled by this program.